+++ src/com/opencms/file/I_CmsResourceType.java
@@ -7,39 +7,39 @@ import java.util.*;
  * for a resource this resource-type is needed.
  * 
  * @author Andreas Schouten
- * @version $Revision: 1.1 $ $Date: 1999/12/03 11:57:10 $
+ * @version $Revision: 1.2 $ $Date: 1999/12/06 09:39:22 $
  */
 public interface I_CmsResourceType {	
 
 	/**
 	 * The resource type-id for a folder.
 	 */
-	public final static long C_FOLDER		= 0;
+	public final static int C_FOLDER		= 0;
 	
 	/**
 	 * The resource type-id for a plain text file.
 	 */
-	public final static long C_TEXTPLAIN	= 1;
+	public final static int C_TEXTPLAIN		= 1;
 	
 	/**
 	 * The resource type-id for a binary file.
 	 */
-	public final static long C_BINARY		= 2;
+	public final static int C_BINARY		= 2;
 	
 	/**
 	 * The resource type-id for a xml base file.
 	 */
-	public final static long C_XMLBASE		= 3;
+	public final static int C_XMLBASE		= 3;
 
 	/**
 	 * The resource type-id for a xml templatefile.
 	 */
-	public final static long C_XMLTEMPLATE	= 4;
+	public final static int C_XMLTEMPLATE	= 4;
 	
 	/**
 	 * The resource type-id for a docloader file.
 	 */
-	public final static long C_DOCLOADER	= 5;
+	public final static int C_DOCLOADER	= 5;
 	
 	/**
 	 * Returns all ResourceTypes in a hashtable. The ResourceType-name is the
